// src/api/models/PaginatedResponseConfigurationSuiteWithWorkerResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ConfigurationSuiteWithWorkerResponse } from './ConfigurationSuiteWithWorkerResponse';
export type PaginatedResponseConfigurationSuiteWithWorkerResponse = {
    content: Array<ConfigurationSuiteWithWorkerResponse>;
    currentPage: number;
    totalPages: number;
    size: number;
    totalElements: number;
};